move card components to a cards subfolder of components
